Add files via upload
The c() function is used to combine the elements 16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22 into a single vector assignment2. The function myMean takes in the vector assignment2 as an argument, calculates the sum of all elements in the vector using the sum() function and divides that sum by the length of the vector using the length() function to find the mean. The mean is returned by the return() function.

You can use the function by calling it by myMean(assignment2) and it will return the mean of the elements in the vector assignment2. R also has a built-in function called mean() that can be used to calculate the average of a vector, so that function could be used instead of creating a custom function.  The built in function can be called by mean() assignment2